@@ -13,14 +13,14 @@ Gem::Specification.new do |spec|
 
   spec.homepage = 'https://github.com/rightscale/attributor'
   spec.license = 'MIT'
-  spec.required_ruby_version = '>=2.1'
+  spec.required_ruby_version = '>=2.7'
 
   spec.require_paths = ['lib']
   spec.files         = `git ls-files -z`.split("\x0")
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
 
   spec.add_runtime_dependency('hashie', ['~> 3'])
-  spec.add_runtime_dependency('randexp', ['~> 0'])
+  spec.add_runtime_dependency('faker', [' >= 3.2'])
   spec.add_runtime_dependency('activesupport', ['>= 3'])
 
   spec.add_development_dependency 'rspec', '~> 3'
@@ -34,6 +34,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency('guard', ['~> 2'])
   spec.add_development_dependency('guard-rspec', ['~> 4'])
   spec.add_development_dependency('pry')
+
   if RUBY_PLATFORM !~ /java/
     spec.add_development_dependency('pry-byebug')
     spec.add_development_dependency('pry-stack_explorer')
