@@ -1,6 +1,6 @@
 class HashWithModel < Attributor::Hash
   keys do
-    key :name, String, default: 'Turkey McDucken', description: 'Turducken name', example: Randgen.name
+    key :name, String, default: 'Turkey McDucken', description: 'Turducken name', example: proc { Faker::Name.first_name }
     key :chicken, Chicken
   end
 end
